requirements.txt
dspy-ai>=2.4.0
langgraph>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
click>=8.1.7
rich>=13.7.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
markdown>=3.5.0
mkdocs>=1.5.0
mkdocs-material>=9.4.0

# Optional: Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0

# Optional: Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Optional: Advanced NLP
transformers>=4.35.0
torch>=2.1.0  # CPU version, adjust for GPU if needed

# Optional: Vector databases (if needed)
# chromadb>=0.4.0
# faiss-cpu>=1.7.4

# Optional: Monitoring and logging
loguru>=0.7.0
rich>=13.6.0
